skip tests on nanoserver
